Insertion of measurements should not fail
There’s the underlying assumption that we are an append-only database, but, we want large queries to have a good chance to succeed, even if some data is already in the database.

`ON CONFLICT DO NOTHING` should help to ensure this. All non-conflicting data entries will still be inserted.

Use actual number of affected rows to get correct statistics.

Signed-off-by: Stefan Marr <git@stefan-marr.de>