Implement timeline plots for use directly on the change report
This encodes the parameters of the benchmark on the button, and uses project name and baseline and change commit ids from the meta data of the page.

To distinguish the baseline and change branches, we request them first based on the commit ids.

The response to the UI is already prepared for display.

The plots are generated with uPlot, a lightweight canvas-based library. This seems more scalable than plot.ly.

To make TypeScript happy about uPlot, added path mappings, which makes the code work in the browser, too.

Signed-off-by: Stefan Marr <git@stefan-marr.de>