Introduce a timed cache mechanism to allow caching query results in dashboard
The idea is that even when we actively invalidate the cache because its backing store changed, we hold on to the old results for a given time.

The reason for this is the self-tracking, which updates the database, and thus, invalidates the cache immediately.
But, realistically, we don’t need immediately correct results.

The standard setting for the invalidation delay is 0ms, thus, we invalidate immediately. This is for testing.

The delay used for normal operations is 5 minutes.

Signed-off-by: Stefan Marr <git@stefan-marr.de>