Add tests and a migration for the added expo in Trial’s unique constraint
This completes the changes started in #115.

While just a theoretical issue, including the expId into the unique constraint prevents issues with unrelated experiments not being able to record their trial information.

This is in practice very unlikely to be an issue since the start time is part of the unique constraint. Though, for benchmarking it was noticed, and felt wrong not to fix it.

Signed-off-by: Stefan Marr <git@stefan-marr.de>