Use EJS template engine for the project page
EJS has support for template includes which is useful for more complex things and code reuse.
So, I am considering using it when moving away from R.

EJS is a pretty lightweight shim over JavaScript, which makes it very powerful.

The syntax is going to take some getting used to…
I am using {% %} as delimiters.

Signed-off-by: Stefan Marr <git@stefan-marr.de>