Add calculateSummaryOfChangeSummaries()
This one determines the summary statistics over the summary statistics.

We take the median of the change median, the min and the max.
This means, we also stop using the geomean.
We just treat the changes as a distribution of values, which we
try to characterize with min, max, median.