Remove use of expId, and instead rely on commit ids
There can be multiple experiments for the same commit id.
This is generally a little problematic, but since we designed the user interface around commit ids, I think this is something we need to handle.
I would assume that multiple experiments do not introduce extra changes, and should just be treated like multiple trials from the same experiment.

This is most relevant for the warmup data visualization and the profiles.
Profiles should be easy enough. We will simply show a list of available profiles.
For warmup data plotting, well, currently everything is going to be shown on the same plot.

Not sure how well that works, but it’s a rare case, so, we can solve it of there are actually problems showing up.

We also now need just the runId in the button info for profile and warmup buttons. We simply use the full commit hashes from the page header where we already have them.